requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Environment
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0
PyJWT>=2.8.0

# Chess
python-chess>=1.999
berserk>=0.13.0

# Image rendering
pillow>=10.2.0
qrcode[pil]>=7.4.2
